refactor: comprehensive code cleanup and architecture improvements
- Extract DeleteConfirmModal to shared component, breaking circular dependency
- Fix multi-select functionality in TaskBoardView by forwarding props to DraggableTaskCard
- Remove unused imports across multiple components (useDrag, CheckSquare, etc.)
- Remove dead code: unused state variables, helper functions, and constants
- Replace duplicate debounce implementation with shared utility
- Tighten DnD item typing for better type safety
- Update all import paths to use shared DeleteConfirmModal component

These changes reduce bundle size, improve code maintainability, and follow the
project's "remove dead code immediately" principle while maintaining full functionality.
